# docs/archive/openapi/bff-dff3ea2.openapi.json
{
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "schemas": {
...
      "Problem": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "title": { "type": "string" },
          "traceId": { "type": "string" },
          "detail": { "type": "string" },
          "type": { "type": "string" },
          "status": { "type": "integer" },
          "code": { "type": "string" }
        },
        "required": ["type", "title", "status", "code", "detail", "traceId"]
      }
    }
  },
  "openapi": "3.1.1",
  "servers": [{ "url": "/" }],
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthenticationController_login_post__api_v1_auth_login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": ["loginType", "loginIdentifier", "password"],
                "properties": {
                  "deviceName": { "type": ["string", "null"] },
                  "loginIdentifier": { "type": "string", "minLength": 1 },
                  "loginType": { "type": "string", "minLength": 1 },
                  "password": { "type": "string", "minLength": 1 }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "accessToken": { "type": ["string", "null"] },
                    "accountId": { "type": ["string", "null"], "format": "uuid" },
                    "authState": { "type": ["string", "null"] },
                    "refreshToken": { "type": ["string", "null"] },
                    "sessionId": { "type": ["string", "null"], "format": "uuid" }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Problem response",
            "content": {
              "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "AuthenticationController_logout_post__api_v1_auth_logout",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": ["revoked"],
                  "properties": { "revoked": { "type": "boolean" } }
                }
              }
            }
          },
          "default": {
            "description": "Problem response",
            "content": {
              "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "AuthenticationController_refresh_post__api_v1_auth_refresh",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": ["sessionId", "refreshToken"],
                "properties": {
                  "refreshToken": { "type": "string", "minLength": 1 },
                  "sessionId": { "type": "string", "format": "uuid" }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "accessToken": { "type": ["string", "null"] },
                    "expiresAt": { "type": ["string", "null"], "format": "date-time" },
                    "refreshToken": { "type": ["string", "null"] }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Problem response",
            "content": {
              "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
            }
          }
        }
      }
    },
...
    "/api/v1/me": {
      "get": {
        "operationId": "MeController_current_get__api_v1_me",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "accountId",
                    "personId",
                    "displayName",
                    "loginType",
                    "loginIdentifier",
                    "institutionalEmail",
                    "roleCode",
                    "availableSystemCodes"
                  ],
                  "properties": {
                    "accountId": { "type": "string", "format": "uuid" },
                    "availableSystemCodes": { "items": { "type": "string" }, "type": "array" },
                    "displayName": { "type": "string" },
                    "institutionalEmail": { "type": "string" },
                    "loginIdentifier": { "type": "string" },
                    "loginType": { "type": "string" },
                    "personId": { "type": "string", "format": "uuid" },
                    "roleCode": { "type": "string" }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Problem response",
            "content": {
              "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
            }
          }
        }
      }
    },
    "/api/v1/me/password": {
      "post": {
        "operationId": "MePasswordController_changePassword_post__api_v1_me_password",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": ["newPassword"],
                "properties": {
                  "currentPassword": { "type": ["string", "null"] },
                  "newPassword": { "type": "string", "minLength": 1 }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "accessToken": { "type": ["string", "null"] },
                    "accountId": { "type": ["string", "null"], "format": "uuid" },
                    "authState": { "type": ["string", "null"] },
                    "refreshToken": { "type": ["string", "null"] },
                    "sessionId": { "type": ["string", "null"], "format": "uuid" }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Problem response",
            "content": {
              "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } }
            }
          }
        }
      }
    },
...
  },
  "info": { "title": "Estoma Platform BFF API", "version": "v1" }
}